Type of Task 6 was fixed

+++ src/by/bsuir/WT/lab1/task6/logic/Task6.java
@@ -1,8 +1,8 @@
 package by.bsuir.WT.lab1.task6.logic;
 
 public class Task6 {
-    public static int[][] makeSquareMatrix(int[] numbers, int size){
-        int[][] matrix = new int[size][];
+    public static double[][] makeSquareMatrix(double[] numbers, int size){
+        double[][] matrix = new double[size][];
         for (int i = 0; i < size; ++i){
             matrix[i] = getShiftedArray(numbers, size, i);
         }
@@ -10,8 +10,8 @@ public class Task6 {
         return matrix;
     }
 
-    private static int[] getShiftedArray(int[] numbers, int size, int shiftDegree){
-        int[] newNumbers = new int[size];
+    private static double[] getShiftedArray(double[] numbers, int size, int shiftDegree){
+        double[] newNumbers = new double[size];
         for (int i = 0; i < size; ++i) {
             newNumbers[i] =  numbers[(i + shiftDegree) % size];
         }
